requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-filter==23.5
djangorestframework==3.14.0
idna==3.6
iniconfig==2.0.0
Markdown==3.5.1
packaging==23.2
pluggy==1.3.0
psycopg==3.1.14
psycopg2-binary==2.9.9
pytest==7.4.3
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
